move union subtypes into separate files
