+++ mod.ts
@@ -2,6 +2,7 @@ export * from "./array.ts";
 export * from "./bool.ts";
 export * from "./common.ts";
 export * from "./compact.ts";
+export * from "./comparable_value_union.ts";
 export * from "./dummy.ts";
 export * from "./implicit_init_num_enum.ts";
 export * from "./int.ts";
@@ -9,5 +10,6 @@ export * from "./option.ts";
 export * from "./record.ts";
 export * from "./result.ts";
 export * from "./str.ts";
+export * from "./tagged_union.ts";
 export * from "./tuple.ts";
 export * from "./union.ts";
