+++ result.test.ts
@@ -28,7 +28,7 @@ Deno.test("Results", () => {
       case 5: {
         const c = new s.Result(
           undefined as any,
-          new s.Err(Err, new s.Record(["a", s.str])),
+          new s.Err(Err, s.record(["a", s.str])),
         );
         const d = c.decode(bytes);
         asserts.assert(d instanceof Err);
@@ -41,7 +41,7 @@ Deno.test("Results", () => {
           undefined as any,
           new s.Err(
             Err,
-            new s.Record(
+            s.record(
               [0, s.str],
               [1, s.str],
             ),
@@ -56,7 +56,7 @@ Deno.test("Results", () => {
       case 7: {
         const c = new s.Result(
           undefined as any,
-          new s.Err(Err, new s.Record(["x", s.str])),
+          new s.Err(Err, s.record(["x", s.str])),
         );
         const d = c.decode(bytes);
         asserts.assert(d instanceof Err);
